fix(kc-saml-plugin): fixed IDP creation issue, enum values and removed kc lib dependency (#7238)
* fix(kc-saml-plugin): remove kc lib dependency - wip


* fix(kc-saml-plugin): removing dependency of kc lib wip


* fix(kc-saml-plugin): removing dependency of kc lib wip


* fix(config-api) keycloak lib conflict issue


* fix(config-api) keycloak lib conflict issue


* fix(config-api) keycloak lib conflict issue


* fix(config-api) keycloak lib conflict issue


* fix(config-api) keycloak lib conflict issue


* fix(kc-config-api): removing kc lib dependency - wip


* fix(kc-config-api): removing kc lib dependency - wip


* fix(kc-config-api): removing kc lib dependency - wip


* feat(kc-config-api): removed kc lib dependency


* feat(kc-config-api): removed kc lib dependency


* feat(kc-config-api): removed kc lib dependency


* feat(kc-config-api): removed kc lib dependency


* feat(kc-config-api): sync with main


* feat(kc-saml-plugin): sonar code issue fixed


---------

+++ jans-config-api/plugins/kc-saml-plugin/src/main/java/io/jans/configapi/plugin/saml/service/IdpService.java
@@ -6,11 +6,14 @@
 
 package io.jans.configapi.plugin.saml.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import io.jans.as.common.service.OrganizationService;
+import io.jans.configapi.core.util.DataUtil;
 import io.jans.configapi.util.AuthUtil;
+
 import io.jans.configapi.plugin.saml.client.IdpClientFactory;
-import io.jans.configapi.plugin.saml.mapper.IdentityProviderMapper;
 import io.jans.configapi.plugin.saml.model.IdentityProvider;
+import io.jans.configapi.plugin.saml.util.Constants;
 import io.jans.model.SearchRequest;
 import io.jans.orm.PersistenceEntryManager;
 import io.jans.orm.model.PagedResult;
@@ -18,7 +21,6 @@ import io.jans.util.exception.InvalidAttributeException;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.ws.rs.core.Response;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -27,8 +29,6 @@ import java.util.Map;
 
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
-
-import org.keycloak.representations.idm.IdentityProviderRepresentation;
 
 @ApplicationScoped
 public class IdpService {
@@ -52,9 +52,6 @@ public class IdpService {
     KeycloakService keycloakService;
 
     @Inject
-    IdentityProviderMapper identityProviderMapper;
-
-    @Inject
     IdpClientFactory idpClientFactory;
 
     @Inject
@@ -68,8 +65,8 @@ public class IdpService {
         return samlConfigService.getSpMetadataUrl(realm, name);
     }
 
-    public List<IdentityProvider> getAllIdentityProviders() {
-        return this.identityProviderService.getAllIdentityProvider(0);
+    public List<IdentityProvider> getAllIdentityProviders(String realmName) throws IOException {
+        return keycloakService.findAllIdentityProviders(realmName);
     }
 
     public IdentityProvider getIdentityProviderByInum(String inum) {
@@ -84,6 +81,15 @@ public class IdpService {
         return identityProviderService.getIdentityProvider(searchRequest);
     }
 
+    public List<IdentityProvider> getAllIdp(String realmName) throws IOException {
+        log.info("Fetch all IDP from realm:{}", realmName);
+        if (StringUtils.isBlank(realmName)) {
+            realmName = Constants.REALM_MASTER;
+        }
+        return keycloakService.findAllIdentityProviders(realmName);
+
+    }
+
     public IdentityProvider createSamlIdentityProvider(IdentityProvider identityProvider, InputStream idpMetadataStream)
             throws IOException {
         log.info(
@@ -95,41 +101,19 @@ public class IdpService {
             throw new InvalidAttributeException("IdentityProvider object is null!!!");
         }
 
-        if (idpMetadataStream != null && idpMetadataStream.available() > 0) {
-            // validate metadata and set in config
-            Map<String, String> config = validateSamlMetadata(identityProvider.getRealm(), idpMetadataStream);
-            log.info("Validated metadata to create IDP - config:{}", config);
-            identityProvider.setConfig(config);
-        } else {
-            // ensure individual metadata elements are present
-            boolean validConfig = validateIdpMetadataElements(identityProvider);
-            log.info("Is metadata individual elements for IDP creation present:{}", validConfig);
-        }
+        // Set Inum
+        String inum = identityProviderService.generateInumForIdentityProvider();
+        identityProvider.setInum(inum);
+        identityProvider.setDn(identityProviderService.getDnForIdentityProvider(inum));
+
+        // common code
+        identityProvider = processIdentityProvider(identityProvider, idpMetadataStream, false);
+        log.debug("Create IdentityProvider identityProvider:{}", identityProvider);
 
         // Create IDP in Jans DB
-        log.debug("Create IdentityProvider identityProvider:{})", identityProvider);
         identityProviderService.addSamlIdentityProvider(identityProvider, idpMetadataStream);
-        log.debug("Created IdentityProvider in Jans DB -  identityProvider:{})", identityProvider);
+        log.debug("Created IdentityProvider in Jans DB -  identityProvider:{}", identityProvider);
 
-        if (samlConfigService.isSamlEnabled()) {
-            // Create IDP in KC
-            IdentityProviderRepresentation kcIdp = this.convertToIdentityProviderRepresentation(identityProvider);
-            log.debug("converted kcIdp:{}", kcIdp);
-
-            log.debug("IDP Service idpMetadataStream:{}, identityProvider.getRealm():{}", idpMetadataStream,
-                    identityProvider.getRealm());
-            kcIdp = keycloakService.createIdentityProvider(identityProvider.getRealm(), kcIdp);
-            log.debug("Newly created kcIdp:{}", kcIdp);
-            identityProvider = this.convertToIdentityProvider(identityProvider, kcIdp);
-            log.debug("Final created identityProvider:{}", identityProvider);
-
-            // set KC SP MetadataURL name
-            if (identityProvider != null) {
-                String spMetadataUrl = getSpMetadataUrl(identityProvider.getRealm(), identityProvider.getName());
-                log.info(" Setting KC SP Metadata URL - spMetadataUrl:{} ", spMetadataUrl);
-                identityProvider.setSpMetaDataURL(spMetadataUrl);
-            }
-        }
         return identityProvider;
     }
 
@@ -144,67 +128,46 @@ public class IdpService {
             throw new InvalidAttributeException("IdentityProvider object for update is null!!!");
         }
 
-        if (idpMetadataStream != null && idpMetadataStream.available() > 0) {
-            // validate metadata and set in config
-            Map<String, String> config = validateSamlMetadata(identityProvider.getRealm(), idpMetadataStream);
-            log.debug("Validated metadata to update IDP - config:{}", config);
-            identityProvider.setConfig(config);
-        } else {
-            // ensure individual metadata elements are present
-            boolean validConfig = validateIdpMetadataElements(identityProvider);
-            log.info("Is metadata individual for update elements present:{}", validConfig);
-        }
-
-        // validate metadata and set in config
-        Map<String, String> config = validateSamlMetadata(identityProvider.getRealm(), idpMetadataStream);
-        log.debug("Validated metadata to update config:{}", config);
-        identityProvider.setConfig(config);
+        // common code
+        identityProvider = processIdentityProvider(identityProvider, idpMetadataStream, true);
+        log.debug("Update IdentityProvider identityProvider:{}", identityProvider);
 
         // Update IDP in Jans DB
         updateIdentityProvider(identityProvider);
-        log.debug("Updated IdentityProvider dentityProvider:{}, , identityProvider.getRealm():{})", identityProvider,
-                identityProvider.getRealm());
-
-        if (samlConfigService.isSamlEnabled()) {
-            // Update IDP in KC
-            IdentityProviderRepresentation kcIdp = this.convertToIdentityProviderRepresentation(identityProvider);
-            log.debug("converted kcIdp:{}", kcIdp);
-
-            kcIdp = keycloakService.updateIdentityProvider(identityProvider.getRealm(), kcIdp);
-            log.debug("Updated kcIdp:{}", kcIdp);
-            identityProvider = this.convertToIdentityProvider(identityProvider, kcIdp);
-
-            // set KC SP MetadataURL name
-            if (identityProvider != null) {
-                String spMetadataUrl = getSpMetadataUrl(identityProvider.getRealm(), identityProvider.getName());
-                log.info(" Updating KC SP Metadata URL - spMetadataUrl:{} ", spMetadataUrl);
-                identityProvider.setSpMetaDataURL(spMetadataUrl);
-            }
-        }
+        log.info("Updated IdentityProvider - identityProvider:{}", identityProvider);
         return identityProvider;
     }
 
     public void deleteIdentityProvider(IdentityProvider identityProvider) {
+        boolean status = false;
+        log.info("Delete dentityProvider:{}, samlConfigService.isSamlEnabled():{}", identityProvider,
+                samlConfigService.isSamlEnabled());
+        // validate
+        if (identityProvider == null) {
+            throw new InvalidAttributeException("IdentityProvider object for delete is null!!!");
+        }
 
         if (samlConfigService.isSamlEnabled()) {
             // Delete IDP in KC
-            keycloakService.deleteIdentityProvider(identityProvider.getRealm(), identityProvider.getName());
+            status = keycloakService.deleteIdentityProvider(identityProvider.getRealm(), identityProvider.getName());
         }
+        log.info("Delete IDP status:{},)", status);
         // Delete in Jans DB
-        identityProviderService.removeIdentityProvider(identityProvider);
+        if (status) {
+            identityProviderService.removeIdentityProvider(identityProvider);
+        }
     }
 
     public void processUnprocessedIdpMetadataFiles() {
         identityProviderService.processUnprocessedIdpMetadataFiles();
     }
 
-    public Response getSpMetadata(IdentityProvider identityProvider) {
-        Response response = null;
+    public String getSpMetadata(IdentityProvider identityProvider) throws JsonProcessingException {
+
         if (identityProvider == null) {
-            return response;
+            throw new InvalidAttributeException("IdentityProvider object is null!!!");
         }
-        return idpClientFactory
-                .getSpMetadata(getSpMetadataUrl(identityProvider.getRealm(), identityProvider.getName()));
+        return keycloakService.getSpMetadata(identityProvider.getRealm(), identityProvider.getName());
 
     }
 
@@ -213,97 +176,138 @@ public class IdpService {
 
         // Update IDP in Jans DB
         identityProviderService.updateIdentityProvider(identityProvider);
-        log.debug("Updated IdentityProvider in Jans DB -  identityProvider:{})", identityProvider);
+        log.debug("Updated IdentityProvider in Jans DB -  identityProvider:{}", identityProvider);
 
         return identityProvider;
     }
 
-    private Map<String, String> validateSamlMetadata(String realmName, InputStream idpMetadataStream) {
-        return keycloakService.validateSamlMetadata(realmName, idpMetadataStream);
-    }
-
-    private IdentityProvider convertToIdentityProvider(IdentityProvider identityProvider,
-            IdentityProviderRepresentation kcIdp) {
-        log.debug("identityProvider:{}, kcIdp:{}", identityProvider, kcIdp);
-
-        IdentityProvider idp = this.convertToIdentityProvider(kcIdp);
-        log.info("convertToIdentityProvider - idp:{}", idp);
-
-        if (idp != null && identityProvider != null) {
-            idp.setRealm(identityProvider.getRealm());
-            idp.setSpMetaDataFN(identityProvider.getSpMetaDataFN());
-            idp.setSpMetaDataURL(identityProvider.getSpMetaDataURL());
-            idp.setSpMetaDataLocation(identityProvider.getSpMetaDataLocation());
-            idp.setIdpMetaDataFN(identityProvider.getIdpMetaDataFN());
-            idp.setIdpMetaDataLocation(identityProvider.getIdpMetaDataLocation());
-            idp.setIdpMetaDataURL(identityProvider.getIdpMetaDataURL());
-            idp.setStatus(identityProvider.getStatus());
-            idp.setValidationStatus(identityProvider.getValidationStatus());
-            idp.setValidationLog(identityProvider.getValidationLog());
+    private IdentityProvider setSamlIdentityProviderDefaultValue(IdentityProvider identityProvider, boolean update) {
+        log.info("Setting default value for identityProvider:{}, update:{}", identityProvider, update);
+        if (identityProvider == null) {
+            return identityProvider;
         }
 
-        return idp;
+        // Set default Realm in-case null
+        if (StringUtils.isBlank(identityProvider.getRealm())) {
+            identityProvider.setRealm(Constants.REALM_MASTER);
+        }
+
+        if (StringUtils.isBlank(identityProvider.getProviderId())) {
+            identityProvider.setProviderId(Constants.SAML);
+        }
+        return identityProvider;
     }
 
-    private IdentityProvider convertToIdentityProvider(IdentityProviderRepresentation kcIdp) {
-        log.debug("kcIdp:{}", kcIdp);
-        IdentityProvider idp = null;
-        if (kcIdp == null) {
-            return idp;
-        }
-        idp = identityProviderMapper.kcIdentityProviderToIdentityProvider(kcIdp);
-        log.info("convertToIdentityProvider - idp:{}", idp);
+    private IdentityProvider processIdentityProvider(IdentityProvider identityProvider, InputStream idpMetadataStream,
+            boolean isUpdate) throws IOException {
+        log.info("Common processing for identityProvider:{}, idpMetadataStream:{}, isUpdate:{}", identityProvider,
+                idpMetadataStream, isUpdate);
 
-        return idp;
+        if (identityProvider == null) {
+            return identityProvider;
+        }
+
+        // Set default Value for SAML IDP
+        setSamlIdentityProviderDefaultValue(identityProvider, isUpdate);
+
+        // SAML IDP Metadata handling
+        if (idpMetadataStream != null && idpMetadataStream.available() > 0) {
+            Map<String, String> config = validateSamlMetadata(identityProvider.getProviderId(),
+                    identityProvider.getRealm(), idpMetadataStream);
+            log.debug("Validated metadata to create IDP - config:{}", config);
+            populateIdpMetadataElements(identityProvider, config);
+        }
+
+        // ensure individual metadata elements are present
+        boolean validConfig = validateIdpMetadataElements(identityProvider);
+        log.info("Is metadata individual elements for IDP creation present:{}", validConfig);
+
+        if (samlConfigService.isSamlEnabled()) {
+            // Create IDP in KC
+            log.info("Create/Update IDP Service idpMetadataStream:{}, identityProvider.getRealm():{}",
+                    idpMetadataStream, identityProvider.getRealm());
+            identityProvider = keycloakService.createUpdateIdentityProvider(identityProvider.getRealm(), isUpdate,
+                    identityProvider);
+
+            log.info("Newly created identityProvider:{}", identityProvider);
+
+            // set KC SP MetadataURL name
+            if (identityProvider != null) {
+                String spMetadataUrl = getSpMetadataUrl(identityProvider.getRealm(), identityProvider.getName());
+                log.debug(" Setting KC SP Metadata URL - spMetadataUrl:{} ", spMetadataUrl);
+                identityProvider.setSpMetaDataURL(spMetadataUrl);
+            }
+        }
+        return identityProvider;
     }
 
-    private IdentityProviderRepresentation convertToIdentityProviderRepresentation(IdentityProvider idp) {
-        log.info("idp:{}", idp);
-        IdentityProviderRepresentation kcIdp = null;
-        if (idp == null) {
-            return kcIdp;
-        }
-        kcIdp = identityProviderMapper.identityProviderToKCIdentityProvider(idp);
-        log.debug("convert IdentityProviderRepresentation - kcIdp:{}", kcIdp);
-
-        log.trace(
-                "convert IDP data kcIdp.getAlias():{}, kcIdp.getInternalId():{}, kcIdp.getProviderId():{}, kcIdp.getConfig():{}, kcIdp.isEnabled():{}, kcIdp.isLinkOnly():{}, kcIdp.isStoreToken():{},kcIdp.getFirstBrokerLoginFlowAlias():{}, kcIdp.getPostBrokerLoginFlowAlias():{},kcIdp.isTrustEmail():{}",
-                kcIdp.getAlias(), kcIdp.getInternalId(), kcIdp.getProviderId(), kcIdp.getConfig(), kcIdp.isEnabled(),
-                kcIdp.isLinkOnly(), kcIdp.isStoreToken(), kcIdp.getFirstBrokerLoginFlowAlias(),
-                kcIdp.getPostBrokerLoginFlowAlias(), kcIdp.isTrustEmail());
-
-        return kcIdp;
+    private Map<String, String> validateSamlMetadata(String prorviderId, String realmName,
+            InputStream idpMetadataStream) throws JsonProcessingException {
+        return keycloakService.importSamlMetadata(prorviderId, realmName, idpMetadataStream);
     }
 
     private boolean validateIdpMetadataElements(IdentityProvider identityProvider) {
-        log.info("identityProvider:{}, samlConfigService.getIdpMetadataMandatoryAttributes():{}", identityProvider, samlConfigService.getIdpMetadataMandatoryAttributes());
+        log.info("identityProvider:{}, samlConfigService.getIdpMetadataMandatoryAttributes():{}", identityProvider,
+                samlConfigService.getIdpMetadataMandatoryAttributes());
         boolean isValid = false;
-        if (identityProvider == null || identityProvider.getConfig() == null
-                || identityProvider.getConfig().isEmpty() || samlConfigService.getIdpMetadataMandatoryAttributes().isEmpty()) {
+        if (identityProvider == null || samlConfigService.getIdpMetadataMandatoryAttributes().isEmpty()) {
             isValid = true;
             return isValid;
         }
 
         List<String> missingElements = null;
-        for(String attribute: samlConfigService.getIdpMetadataMandatoryAttributes()) {
-            log.info("attribute:{}", attribute);
-            if(StringUtils.isBlank(identityProvider.getConfig().get(attribute))){
-                if(missingElements == null) {
+        for (String attribute : samlConfigService.getIdpMetadataMandatoryAttributes()) {
+            log.debug("attribute:{}, getValue(identityProvider, attribute):{}", attribute,
+                    getValue(identityProvider, attribute));
+            if (StringUtils.isBlank(getValue(identityProvider, attribute))) {
+                if (missingElements == null) {
                     missingElements = new ArrayList<>();
                 }
                 missingElements.add(attribute);
-            }           
+            }
         }
-      
+
         log.info("missingElements:{}", missingElements);
 
         if (missingElements != null && !missingElements.isEmpty()) {
             isValid = false;
-            log.error("IDP elements are missing:{}, isValid:{} !", missingElements, isValid);
-            throw new InvalidAttributeException("IDP mandatory attribute missing - "+missingElements+" !!!");
+            log.debug("IDP elements are missing:{}, isValid:{} !", missingElements, isValid);
+            throw new InvalidAttributeException("IDP mandatory attribute missing - " + missingElements + " !!!");
         }
         isValid = true;
         log.info("validateIdpMetadataElements - isValid:{}", isValid);
         return isValid;
     }
+
+    private IdentityProvider populateIdpMetadataElements(IdentityProvider identityProvider,
+            Map<String, String> config) {
+        log.info("identityProvider:{}, config:{}, samlConfigService.getKcSamlConfig():{}", identityProvider, config,
+                samlConfigService.getKcSamlConfig());
+
+        if (identityProvider == null || config == null || samlConfigService.getKcSamlConfig().isEmpty()) {
+            return identityProvider;
+        }
+
+        for (String attribute : samlConfigService.getKcSamlConfig()) {
+            log.trace("attribute:{}, config.get(attribute):{}", attribute,
+                    config.get(attribute));
+                DataUtil.invokeReflectionSetter(identityProvider, attribute, config.get(attribute));        
+        }
+
+        log.info("validateIdpMetadataElements - identityProvider:{}", identityProvider);
+        return identityProvider;
+    }
+
+    private String getValue(IdentityProvider identityProvider, String property) {
+        log.debug("Get Field Value - identityProvider:{}, property:{}", identityProvider, property);
+        String value = null;
+        try {
+            value = (String) DataUtil.getValue(identityProvider, property);
+            log.debug("Field Value - property:{}, value:{}", property, value);
+        } catch (Exception ex) {
+            log.error("Error while getting value of config ", ex);
+        }
+        return value;
+    }
+
 }
